Breakout the console code so it can be reused.
No functional change.
