+++ hello_world/console.h
@@ -1,0 +1,5 @@
+void potato_uart_init(void);
+int getchar(void);
+void putchar(unsigned char c);
+void putstr(const char *str, unsigned long len);
+size_t strlen(const char *s);
